upgrade patient filters
